requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
redis>=4.2.0rc1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beanie==1.26.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
fastapi==0.112.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lazy-model==0.2.0
magic-filter==1.0.12
motor==3.5.1
multidict==6.0.5
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.37.2
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
yarl==1.9.4
